Replace unneeded classes with simple functions
